+++ src/compiler/translator/VariableInfo.cpp
@@ -55,8 +55,7 @@ void ExpandVariable(const ShaderVariable &variable,
     {
         if (variable.isArray())
         {
-            for (unsigned int elementIndex = 0; elementIndex < variable.elementCount();
-                 elementIndex++)
+            for (size_t elementIndex = 0; elementIndex < variable.elementCount(); elementIndex++)
             {
                 std::string lname = name + ::ArrayString(elementIndex);
                 std::string lmappedName = mappedName + ::ArrayString(elementIndex);
